announcements/announcements/requirements.txt
requests
Werkzeug==2.0.3
flask==2.1.3
flask-lambda-python36
boto3